+++ tools/cli/air-trace/src/stats/log_data.rs
