+++ tools/cli/air-trace/src/stats/report.rs
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-use crate::logs::{LogKey, LogRecord, Message};
+use super::log_data::{LogKey, LogRecord, Message};
 use crate::utils::parse_tracing_duration;
 
 use std::{collections::HashMap, time::Duration};
